requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
asyncpg
pydantic
python-dotenv
aiohttp
openai
imagehash
elasticsearch
